chore(web) : validate schema 화면별 분리 #180
