+++ apps/web/src/validate/order-form-schema.ts
@@ -81,102 +81,3 @@ export const formSchema = z.object({
 });
 
 export type FormSchema = z.infer<typeof formSchema>;
-
-/**
- * 내 정보수정 유효성 검증
- */
-export const customerFormSchema = z.object({
-  name: z // 고객 이름
-    .string({
-      required_error: '이름을 입력 바랍니다.',
-    })
-    .min(minNum, { message: `이름은 ${minMsg}` })
-    .max(5, { message: `이름은 ${maxMsg}` })
-    .regex(/^[가-힣+$]/, { message: `이름은 ${regExpMsg}` }),
-  id: z // 아이디
-    .string({
-      required_error: '아이디를 입력 바랍니다.',
-    })
-    .trim()
-    .min(5, '아이디는 최소 5자 이상 입력 바랍니다.')
-    .max(15, '아이디는 최대 15자 이내로 입력 바랍니다.')
-    .regex(/^[A-Za-z0-9]{5,15}$/, {
-      message: '아이디는 영문 대소문자 최소 5~15글자 입력해야 합니다.',
-    }),
-  mobileNumber: z // 전화번호
-    .string({ required_error: '전화번호를 입력 바랍니다.' })
-    .regex(/^01[016789]-?\d{3,4}-?\d{4}$/, { message: '유효한 휴대번호 양식이 아닙니다.' }),
-  createdAt: z // 등록일시
-    .string(),
-});
-
-export type CustomerFormSchema = z.infer<typeof customerFormSchema>;
-
-/**
- * 비밀번호 변경 유효성 검증
- */
-const PASSWORD_VALIDATION = {
-  min: {
-    value: 10,
-    message: '비밀번호는 10자 이상입니다.',
-  },
-  max: {
-    value: 30,
-    message: '비밀번호는 30자 이내로 작성 바랍니다.',
-  },
-  regex: {
-    value: /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_])[^\s]{10,20}$/g,
-    message:
-      '비밀번호는 영문 대소문자 + 숫자 + 특수문자 포함, 10~20자리, 공백 제외하여 입력 가능합니다.',
-  },
-  required: {
-    message: '비밀번호를 입력하세요.',
-  },
-};
-
-export const passwordModifyFormSchema = z
-  .object({
-    password: z // 현재 비밀번호
-      .string({ required_error: '현재 비밀번호를 입력 바랍니다.' })
-      .trim()
-      .min(PASSWORD_VALIDATION.min.value, { message: PASSWORD_VALIDATION.min.message })
-      .max(PASSWORD_VALIDATION.max.value, { message: PASSWORD_VALIDATION.max.message })
-      .regex(PASSWORD_VALIDATION.regex.value, { message: PASSWORD_VALIDATION.regex.message }),
-    passwordModify: z // 새 비밀번호
-      .string({ required_error: '새 비밀번호를 입력 바랍니다.' })
-      .trim()
-      .min(PASSWORD_VALIDATION.min.value, { message: PASSWORD_VALIDATION.min.message })
-      .max(PASSWORD_VALIDATION.max.value, { message: PASSWORD_VALIDATION.max.message })
-      .regex(PASSWORD_VALIDATION.regex.value, { message: PASSWORD_VALIDATION.regex.message }),
-    passwordConfirm: z // 비밀번호 확인
-      .string({ required_error: '확인을 위한 새 비밀번호를 입력 바랍니다.' })
-      .trim()
-      .min(PASSWORD_VALIDATION.min.value, { message: PASSWORD_VALIDATION.min.message })
-      .max(PASSWORD_VALIDATION.max.value, { message: PASSWORD_VALIDATION.max.message })
-      .regex(PASSWORD_VALIDATION.regex.value, { message: PASSWORD_VALIDATION.regex.message }),
-  })
-  .superRefine(({ password, passwordModify }, ctx) => {
-    // 비밀번호 확인 :: 틀리면 오류
-    if (password !== passwordModify) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        path: ['passwordConfirm'],
-        message: '비밀번호가 일치하지 않습니다.',
-      });
-    }
-  });
-
-export type PasswordModifyFormSchema = z.infer<typeof passwordModifyFormSchema>;
-
-/**
- * 주문 취소
- */
-export const orderCancelFormSchema = z.object({
-  canceledReason: z // 취소사유
-    .string({ required_error: '취소사유를 입력 바랍니다.' })
-    .trim()
-    .min(5, { message: '취소사유는 최소 5자 이상 작성해야 합니다.' })
-    .max(300, { message: '취소사유는 최대 300자 이내로 작성해야 합니다.' }),
-});
-
-export type OrderCancelFormSchema = z.infer<typeof orderCancelFormSchema>;
